added allowance and allowance percent controller and service

+++ src/main/java/com/marcosimon/autosurvey/models/NewAllowanceValueDTO.java
@@ -3,10 +3,10 @@ package com.marcosimon.autosurvey.models;
 public record NewAllowanceValueDTO(
         //AllowanceValueDTO allValue
         String allowancePercentageId,
-        Float COLA,
-        Float transportation,
-        Float housing,
-        Float other,
-        Float total
+        Integer COLA,
+        Integer transportation,
+        Integer housing,
+        Integer other,
+        Integer total
 ) {
 }
